Create separate workflow for AddCollisionId Only applied to X task for now
