+++ Analysis/Tasks/PWGHF/AddCollisionId.cxx
@@ -1,0 +1,27 @@
+#include "Framework/runDataProcessing.h"
+#include "Framework/AnalysisTask.h"
+
+namespace o2::aod
+{
+namespace extra
+{
+DECLARE_SOA_INDEX_COLUMN(Collision, collision);
+}
+DECLARE_SOA_TABLE(Colls, "AOD", "COLLSID", o2::aod::extra::CollisionId);
+} // namespace o2::aod
+
+struct AddCollisionId {
+  Produces<o2::aod::Colls> colls;
+  void process(aod::HfCandProng2 const& candidates, aod::Tracks const&)
+  {
+    for (auto& candidate : candidates) {
+      colls(candidate.index0_as<aod::Tracks>().collisionId());
+    }
+  } // process
+};  // struct
+
+WorkflowSpec defineDataProcessing(ConfigContext const&)
+{
+  return WorkflowSpec{
+    adaptAnalysisTask<AddCollisionId>("hf-task-add-collisionId")};
+}
