+++ Analysis/Tasks/PWGHF/taskX.cxx
@@ -14,6 +14,7 @@
 /// \author Gian Michele Innocenti <gian.michele.innocenti@cern.ch>, CERN
 /// \author Rik Spijkers <r.spijkers@students.uu.nl>, Utrecht University
 
+#include "Framework/runDataProcessing.h"
 #include "Framework/AnalysisTask.h"
 #include "Framework/HistogramRegistry.h"
 #include "AnalysisDataModel/HFSecondaryVertex.h"
@@ -30,28 +31,7 @@ void customize(std::vector<o2::framework::ConfigParamSpec>& workflowOptions)
   workflowOptions.push_back(optionDoMC);
 }
 
-#include "Framework/runDataProcessing.h"
-
-namespace o2::aod
-{
-namespace extra
-{
-DECLARE_SOA_INDEX_COLUMN(Collision, collision);
-}
-DECLARE_SOA_TABLE(Colls, "AOD", "COLLSID", o2::aod::extra::CollisionId);
-} // namespace o2::aod
-struct AddCollisionId {
-  Produces<o2::aod::Colls> colls;
-  void process(aod::HfCandProng2 const& candidates, aod::Tracks const&)
-  {
-    for (auto& candidate : candidates) {
-      colls(candidate.index0_as<aod::Tracks>().collisionId());
-    }
-  }
-};
-
 /// X(3872) analysis task
-/// FIXME: Still need to remove track duplication!!!
 struct TaskX {
   HistogramRegistry registry{
     "registry",
